// src/app/dashboard/interfaces/products-get-response.interface.ts
// Generated by https://quicktype.io

export interface ProductGetResponse {
  _id:          string;
  name:         string;
  unit:         string;
  unitPrice:    number;
  descripcion:  string;
  tipoProducto: string;
  igvIndicator: boolean;
  igv:          number;
  __v:          number;
}


export enum Unit {
  Grueso = "grueso",
  Docena = "docena",
  Kg = "kg",
  Unidad = "unidad",
  Metro = "metro",
  Tonelada = "tonelada",
  Mediadocena = "mediadocena",
}

export enum TipoProducto {
  ProductoPrimeraNecesidad = "producto_primera_necesidad",
  MateriaPrima = "materia_prima",
  Consumible = "consumible",
  Tecnologia = "tecnologia",
  Electrodomestico = "electrodomestico",
  Ropa = "ropa",
}