// src/app/dashboard/interfaces/facture-post-response.interface.ts
// Generated by https://quicktype.io

import { User } from "src/app/auth/interfaces/user.interface";
import { Product } from "./product.interface";

// Generated by https://quicktype.io

export interface FacturePostResponse {
  factureState:   string;
  totalPayments:  number;
  totalPriceNet:  number;
  totalPrice:     number;
  totalIgv:       number;
  payMethod:      string;
  user:           User;
  pay:            any[];
  _id:            string;
  amountPayments: number;
  company:        Client;
  detail:         Detail[];
  client:         Client;
  date?: Date;
}

export interface Client {
  name:         string;
  ruc:          string;
  typeDocument: string;
  adress:       string;
  state:        string;
  condition:    string;
  district:     string;
  province:     string;
  departament:  string;
  _id?:         string;
  __v?:         number;
}

export interface Detail {
  _id:             string;
  product:         Product;
  amount:          number;
  totalPrice:      number;
  igvImportDetail: number;
  __v:             number;
}